Moved downscale to utilities
